feat: cleanup

+++ modules/00-prechroot/10-disk.sh
@@ -9,11 +9,11 @@ sgdisk -Z \
 esp=/dev/disk/by-partlabel/esp
 root=/dev/disk/by-partlabel/root
 
-[ -n "$disk_passwd" ] && {
+if [ -n "$disk_passwd" ]; then
     echo -n "$disk_passwd" | cryptsetup luksFormat "$root" -
     cryptsetup open "$root" cryptroot - <<< "$disk_passwd"
     root=/dev/mapper/cryptroot
-}
+fi
 
 sleep 5 # wait /dev/disk/by-partlabel/
 
@@ -26,13 +26,13 @@ UUID=$(blkid "$root" -ovalue -sUUID)	/	ext4	rw,relatime	0 1
 UUID=$(blkid "$esp" -ovalue -sUUID)	/efi	vfat	rw,fmask=0077,dmask=0077,noauto	0 2
 EOF
 
-[ -n "$swapfile" ] && {
+if [ -n "$swapfile" ]; then
     dd if=/dev/zero of=/mnt/swapfile bs=1M count="$swapfile" status=progress
     chmod 600 /mnt/swapfile
     mkswap /mnt/swapfile
     swapon /mnt/swapfile
     echo "/swapfile	none	swap	defaults	0 0" >> /mnt/etc/fstab
-}
+fi
 
 cp -rfTv rootfs/ /mnt/
